Move ExternalLink to academic scope
