+++ vitae/features/researchers/model/academic/__init__.py
@@ -1,0 +1,3 @@
+from .external import ExternalLinks
+
+__all__ = ["ExternalLinks"]
