+++ vitae/features/researchers/model/researcher.py
@@ -5,39 +5,24 @@ import attrs
 from vitae.features.researchers.lib import optional
 from vitae.infra.database import tables
 
-from ._shared import optional
-from .address import Address
+from .academic import ExternalLinks
 from .cv import Curriculum
-from .external import Lattes, Orcid
 from .personal import Person
 from .professional import Address
-
-
-@attrs.frozen
-class ExternalLink:
-    lattes: Lattes
-    orcid: Orcid | None
-
-    @classmethod
-    def from_table(cls, researcher: tables.Researcher):
-        return cls(
-            lattes=Lattes.from_id(researcher.lattes_id),
-            orcid=optional(researcher.orcid, Orcid.from_url),
-        )
 
 
 @attrs.frozen
 class Researcher:
     this: Person
     curriculum: Curriculum
-    links: ExternalLink
+    links: ExternalLinks
     address: Address | None
 
     @classmethod
     def from_table(cls, table: tables.Researcher):
         return cls(
             this=Person.from_table(table),
-            links=ExternalLink.from_table(table),
+            links=ExternalLinks.from_table(table),
             address=optional(table.address, Address.from_table),
             curriculum=Curriculum.from_table(table),
         )
